requirements.txt
pwntools>=4.0
